feat(event-display): separate dat.GUI UI menu
